py2exe/requirements.txt
requests
pyyaml
pyautogui
pillow
pyjwt
schedule